Rename `Updated at` field in tag information
The field `Updated at` implies that it should display the last time the tag content was updated, but instead this displays the last time the tag was interacted with at all (whether it be a use in community-help, tag content update, rename, etc.). This occurs because Mongoose automatically increments the `updatedAt` field anytime the document is modified, which includes tag usage, since showing a tag increments the `uses` property.